+++ docs/bot_reply.json
@@ -108,7 +108,7 @@
         "你知道嗎，凹凹通過操縱股東大會來控制公司決策",
         "你知道嗎，阿豬頭利用內部消息交易股票，違反證券法"
     ],
-    "meow_responses": [
+    "meowResponses": [
         "喵~",
         "喵喵!",
         "喵喵喵~",
